refactor: add RemotePhoto
